Refactor integration test structure
