+++ .github/workflows/integration_test.yaml
@@ -23,7 +23,8 @@ jobs:
             --path ${{secrets.E2E_TESTING_REPO}} \
             --token ${{secrets.E2E_TESTING_TOKEN}} \
             --token-alt ${{secrets.E2E_TESTING_TOKEN_ALT}}
-      - name:
+      # TODO: Remove this
+      - name: SSH for debug
         if: failure()
         uses: mxschmitt/action-tmate@v3
         with:
